requirements.txt
numpy

lxml

spacy

scikit-learn==1.3.0

asyncio

aiohttp

bs4

joblib

flask[assync]

transformers

nltk

torch 
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cu118 

python-dotenv